+++ src/onegov.gazette/onegov/gazette/views/notices.py
@@ -169,6 +169,7 @@ def view_notices(self, request):
         'term': self.term,
         'from_date': self.from_date,
         'to_date': self.to_date,
+        'current_issue': layout.current_issue,
         'orderings': orderings,
         'clear': request.link(self.for_dates(None, None).for_term(None)),
         'new_notice': request.link(self, name='new-notice'),